+++ lib/rubycode/views/adapter.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Index file for adapter view components
+require_relative "adapter/debug_delay"
+require_relative "adapter/debug_request"
+require_relative "adapter/debug_response"